- updated id from int to long;

+++ src/main/java/com/provisa/dao/Operator.java
@@ -18,7 +18,7 @@ public class Operator {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private long id;
 
     @Column
     private String operatorName;
